@@ -4,15 +4,12 @@ import lavalink.client.io.Link
 import lavalink.client.player.LavalinkPlayer
 
 /**
- * Represents a Guild Music Manager, used to get audio from the [TrackScheduler].
- *
- * @author Callum Seabrook
- * @since 1.0
+ * The music manager for a guild.
  */
 class MusicManager(link: Link) {
 
     val player: LavalinkPlayer = link.player
-    val scheduler = TrackScheduler(player)
+    val scheduler: TrackScheduler = TrackScheduler(player)
 
     init {
         player.addListener(scheduler)
