@@ -8,15 +8,18 @@ import com.sedmelluq.discord.lavaplayer.track.AudioTrack
 import me.bardy.bot.command.Command
 import me.bardy.bot.command.CommandContext
 import me.bardy.bot.components.ManagerMap
-import me.bardy.bot.format
-import me.bardy.bot.formatted
-import me.bardy.bot.logger
-import me.bardy.bot.sumBy
+import me.bardy.bot.util.buildEmbed
+import me.bardy.bot.util.color
+import me.bardy.bot.util.description
+import me.bardy.bot.util.format
+import me.bardy.bot.util.formatName
+import me.bardy.bot.util.logger
+import me.bardy.bot.util.title
 import net.dv8tion.jda.api.EmbedBuilder
 import net.dv8tion.jda.api.entities.Member
 import org.springframework.stereotype.Component
+import java.time.Duration
 import kotlin.random.Random
-import kotlin.time.milliseconds
 
 @Component
 class QueueCommand(private val musicManagers: ManagerMap) : Command("q") {
@@ -29,17 +32,17 @@ class QueueCommand(private val musicManagers: ManagerMap) : Command("q") {
 
     private fun paginateQueue(context: CommandContext, pageNumber: Int = 1) {
         val member = context.member ?: return
-
-        val musicManager = musicManagers[context.guild.id]
+        val musicManager = musicManagers.get(context.guild.id)
         val queue = musicManager.scheduler.queue
 
         val paginatedQueue = queue.chunked(10)
-        val embed = EmbedBuilder()
-            .setTitle("Here's what I've got lined up for you in ${context.guild.name}", "https://bot.bardy.me")
-            .setColor(Random.nextInt(RGB_MAX_VALUE))
+        val embed = buildEmbed {
+            title("Here's what I've got lined up for you in ${context.guild.name}", "https://bot.bardy.me")
+            color(Random.nextInt(RGB_MAX_VALUE))
+        }
 
         if (pageNumber == 1) {
-            embed.setDescription("\n\n__What I'm playing now:__\n")
+            embed.description("\n\n__What I'm playing now:__\n")
 
             val nowPlaying = musicManager.player.playingTrack
             if (nowPlaying == null) {
@@ -51,7 +54,7 @@ class QueueCommand(private val musicManagers: ManagerMap) : Command("q") {
                     return
                 }
 
-                embed.appendTrack(nowPlaying, requester, -1)
+                appendTrack(embed, nowPlaying, requester, -1)
             }
 
             if (paginatedQueue.isEmpty()) {
@@ -70,31 +73,30 @@ class QueueCommand(private val musicManagers: ManagerMap) : Command("q") {
                     return
                 }
 
-                embed.appendTrack(it, requester, i + 1)
+                appendTrack(embed, it, requester, i + 1)
             }
         }
 
         if (queue.size > 0) {
             val plural = if (queue.size == 1) "tune" else "tunes"
-            val length = queue.sumBy { it.duration }.milliseconds.format()
+            val length = Duration.ofMillis(queue.sumOf { it.duration }).format()
             embed.appendDescription("\n**${queue.size} banging $plural ready to be played, $length long**")
         }
 
         context.reply(embed.setFooter("Page $pageNumber / ${paginatedQueue.size}", member.user.avatarUrl).build())
     }
 
-    private fun EmbedBuilder.appendTrack(track: AudioTrack, requester: Member, number: Int) {
-        if (number != -1) appendDescription("\n`$number`. ") else appendDescription("\n")
-        appendDescription("""[${track.info.title}](${track.info.uri}) | `${track.duration.milliseconds.format()}`
-            *Who put it on? ${requester.formatted} did!*
+    private fun appendTrack(embed: EmbedBuilder, track: AudioTrack, requester: Member, number: Int) {
+        if (number != -1) embed.appendDescription("\n`$number`. ") else embed.appendDescription("\n")
+        embed.appendDescription("""[${track.info.title}](${track.info.uri}) | `${Duration.ofMillis(track.duration).format()}`
+            *Who put it on? ${requester.formatName()} did!*
 
         """.trimIndent())
     }
 
     companion object {
 
-        const val RGB_MAX_VALUE = 16777216
-
+        private const val RGB_MAX_VALUE = 16777216
         private val LOGGER = logger<QueueCommand>()
     }
 }
