@@ -2,8 +2,8 @@ package me.bardy.bot.commands.music
 
 import me.bardy.bot.command.Command
 import me.bardy.bot.components.ManagerMap
-import me.bardy.bot.logger
 import me.bardy.bot.services.TrackService
+import me.bardy.bot.util.logger
 import org.springframework.stereotype.Component
 
 @Component
@@ -13,7 +13,7 @@ class SkipCommand(
 ) : Command("sk") {
 
     override fun register() = default("skip") {
-        if (musicManagers[it.guild.id].player.playingTrack == null) {
+        if (musicManagers.get(it.guild.id).player.playingTrack == null) {
             it.reply("I can't skip to the next track if there isn't one already playing to skip! You silly goose.")
             return@default
         }
