@@ -11,8 +11,4 @@ import org.springframework.boot.context.properties.ConstructorBinding
  */
 @ConfigurationProperties("bot")
 @ConstructorBinding
-data class BotConfig(
-    val clientId: String,
-    val token: String,
-    val prefix: String
-)
+data class BotConfig(val clientId: String, val token: String, val prefix: String)
