@@ -4,28 +4,19 @@ import com.mojang.brigadier.CommandDispatcher
 import com.mojang.brigadier.ParseResults
 import com.mojang.brigadier.exceptions.CommandSyntaxException
 import me.bardy.bot.command.CommandContext
-import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent
-import net.dv8tion.jda.api.hooks.ListenerAdapter
-import net.dv8tion.jda.api.sharding.ShardManager
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.stereotype.Component
-import javax.annotation.PostConstruct
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 
 @Component
-class MessageListener(
-    private val dispatcher: CommandDispatcher<CommandContext>,
-    private val shardManager: ShardManager
-) : ListenerAdapter() {
+class MessageListener(private val dispatcher: CommandDispatcher<CommandContext>) : BardyBotListener() {
 
-    @PostConstruct
-    fun registerSelf() = shardManager.addEventListener(this)
-
-    override fun onGuildMessageReceived(event: GuildMessageReceivedEvent) {
+    override fun onMessageReceived(event: MessageReceivedEvent) {
         if (!event.message.contentRaw.startsWith('!')) return
         val message = event.message.contentRaw.removePrefix("!")
 
         try {
-            dispatcher.execute(parseCommand(event.buildContext(), message))
+            dispatcher.execute(parseCommand(CommandContext(event.guild, event.channel, event.member, event.message), message))
         } catch (exception: CommandSyntaxException) {
             val exceptionMessage = when (exception.type) {
                 CommandSyntaxException.BUILT_IN_EXCEPTIONS.dispatcherParseException() -> "a parse exception"
@@ -39,6 +30,4 @@ class MessageListener(
 
     @Cacheable("parse_results")
     fun parseCommand(context: CommandContext, command: String): ParseResults<CommandContext> = dispatcher.parse(command, context)
-
-    private fun GuildMessageReceivedEvent.buildContext() = CommandContext(guild, channel, member, message)
 }
